pyproject.toml
[project]
name = "shepherd"
version = "0.0.0"
description = "flock simulation"
authors = [
    { name = "superwhiskers", email = "whiskerdev@protonmail.com" }
]
dependencies = [
    "prophet>=1.1.5",
    "pandas>=2.2.2",
    "numpy>=1.0.0,<2.0.0",
    "plotly>=5.22.0",
    "ortools>=9.10.4067",
    "scikit-learn>=1.5.1",
    "python-ulid>=2.7.0",
    "textual>=0.72.0",
    "mypy>=1.10.1",
]
license = "0BSD"
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
shepherd = "shepherd:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "black>=24.4.2",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/shepherd"]
